+++ datasketches/src/theta/intersection.rs
@@ -40,16 +40,17 @@ impl RawThetaIntersectionPolicy<ThetaEntry> for NoopIntersectionPolicy {
 }
 
 impl ThetaIntersection {
+    /// Creates a new intersection operator with the default seed.
+    #[allow(clippy::new_without_default)]
+    pub fn new() -> Self {
+        Self::with_seed(DEFAULT_UPDATE_SEED)
+    }
+
     /// Creates a new intersection operator for the given `seed`.
-    pub fn new(seed: u64) -> Self {
+    pub fn with_seed(seed: u64) -> Self {
         Self {
             raw: RawThetaIntersection::new(seed, NoopIntersectionPolicy),
         }
-    }
-
-    /// Creates a new intersection operator with the default seed.
-    pub fn new_with_default_seed() -> Self {
-        Self::new(DEFAULT_UPDATE_SEED)
     }
 
     /// Updates the intersection with a given sketch.